AtomRef to atom_index; cluster_indeces preparation

+++ src/neighbourhood_managers/neighbourhood_manager_lammps.hh
@@ -49,7 +49,8 @@ namespace rascal {
 
 
   //----------------------------------------------------------------------------//
-  class NeighbourhoodManagerLammps: public NeighbourhoodManagerBase<NeighbourhoodManagerLammps>
+  class NeighbourhoodManagerLammps:
+    public NeighbourhoodManagerBase<NeighbourhoodManagerLammps>
   {
   public:
     using traits = NeighbourhoodManager_traits<NeighbourhoodManagerLammps>;
@@ -146,9 +147,9 @@ namespace rascal {
 
     // return the index-th neighbour of cluster
     template<size_t Level, size_t Depth>
-    inline size_t get_cluster_neighbour(const ClusterRefBase<Level, Depth>
-					& cluster,
-					size_t index) const {
+    inline int get_cluster_neighbour(const ClusterRefBase<Level, Depth>
+				     & cluster,
+				     size_t index) const {
       static_assert(Level == traits::MaxLevel-1,
                     "this implementation only handles atoms and pairs");
       auto && i_atom_id{cluster.back()};
@@ -156,8 +157,8 @@ namespace rascal {
     }
 
     // return the atom_index of the index-th atom in manager
-    inline size_t get_cluster_neighbour(const Parent & /*cluster*/,
-					size_t index) const {
+    inline int get_cluster_neighbour(const Parent & /*cluster*/,
+				     size_t index) const {
       return this->ilist[index];
     }
