+++ src/neighbourhood_managers/neighbourhood_manager_minimal.hh
@@ -126,7 +126,7 @@ namespace rascal {
 
     // return the index-th neighbour of cluster
     template<size_t Level, size_t Depth>
-    inline size_t get_cluster_neighbour(const ClusterRefBase<Level, Depth>
+    inline int get_cluster_neighbour(const ClusterRefBase<Level, Depth>
 					& cluster,
 					size_t index) const {
       static_assert(Level <= traits::MaxLevel,
@@ -139,8 +139,8 @@ namespace rascal {
 
 
     // return the atom_index of the index-th atom in manager
-    inline size_t get_cluster_neighbour(const Parent & /*cluster */ ,
-			      size_t index) const {
+    inline int get_cluster_neighbour(const Parent & /*cluster */ ,
+				     size_t index) const {
       return this->centers[index].get_index();
     }
